2015/day13/main_test.go
package main

import "testing"

func TestDinner(t *testing.T) {
	t.Run("Part1", func(t *testing.T) {
		given := `Alice would gain 54 happiness units by sitting next to Bob.
Alice would lose 79 happiness units by sitting next to Carol.
Alice would lose 2 happiness units by sitting next to David.
Bob would gain 83 happiness units by sitting next to Alice.
Bob would lose 7 happiness units by sitting next to Carol.
Bob would lose 63 happiness units by sitting next to David.
Carol would lose 62 happiness units by sitting next to Alice.
Carol would gain 60 happiness units by sitting next to Bob.
Carol would gain 55 happiness units by sitting next to David.
David would gain 46 happiness units by sitting next to Alice.
David would lose 7 happiness units by sitting next to Bob.
David would gain 41 happiness units by sitting next to Carol.`

		want := 330
		got := Part1(makeGraph(given))
		if got != want {
			t.Errorf("got '%v' want '%v'", got, want)
		}
	})
}